requirements.txt
altgraph==0.17.4
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
python-dateutil==2.8.2
pywin32-ctypes==0.2.2
six==1.16.0
vtk==9.2.6